requirements.txt
# 動作確認済み。0.1.46+ で ProcessTransport エラー対策（#386 / PR #630）が入る
claude-agent-sdk>=0.1.46
a2a-sdk[http-server]>=0.3.20
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
